pyproject.toml
[project]
name = "sticker-tools"
version = "0.1.0"
description = ""
authors = [
    {name = "vbidula",email = "vadym.bidula@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "pyqt6 (>=6.9.1,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
sticker = "src.sticker_tools.cli_interface:create_sticker"
